Refactor import paths to capabilities instead of api
