+++ test/e2e/consolidated_controller_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/3scale/3scale-operator/pkg/apis"
-	operator "github.com/3scale/3scale-operator/pkg/apis/api/v1alpha1"
+	operator "github.com/3scale/3scale-operator/pkg/apis/capabilities/v1alpha1"
 	framework "github.com/operator-framework/operator-sdk/pkg/test"
 	v12 "k8s.io/api/core/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
