+++ test/e2e/e2eutil/wait.go
@@ -3,7 +3,7 @@ package e2eutil
 import (
 	"context"
 	"fmt"
-	"github.com/3scale/3scale-operator/pkg/apis/api/v1alpha1"
+	"github.com/3scale/3scale-operator/pkg/apis/capabilities/v1alpha1"
 	"github.com/operator-framework/operator-sdk/pkg/test"
 	"k8s.io/apimachinery/pkg/types"
 	"testing"
